+++ src/wfg/native_ui/ui/system/InteractionSystem.java
@@ -46,8 +46,7 @@ public final class InteractionSystem extends BaseSystem {
                 if (!event.isConsumed() && event.isKeyDownEvent() &&
                     event.getEventValue() == listen.shortcut
                 ) {
-                    listen.onShortcutPressed.run(element);
-                    event.consume();
+                    listen.onShortcutPressed.run(element, event);
                     break;
                 }
             }