+++ changelog.md
@@ -3,6 +3,7 @@ Version 0.4.4-beta
 - SortableTable fixes concerning rows.
 - Added a new overload for NativeUiUtils.setAlpha that takes in a raw alpha value [0, 255].
 - Vanilla CutStyle enums are now mapped to the NativeUI CutStyle enums in Button with a setter.
+- InteractionComp onShortcutPresserd has a new handler that passes the event as an arguemnt and no longer consumes it.
 
 Version 0.4.3-beta
 - Added player bright color to UIConstants.